refactor(state): extract simulation state access
